Notebook 1 added
Multiplex ingestion into a bronze table, the key points to be aware of are the use of helper notebooks that are run at the start. Be aware of the path variables, they are slightly different as this was originally run on a databricks notebook. The readStream and writeStream functions are important to take note of. These will be important for ingesting data. We are ingesting streaming data in this project.