added notebook 4
Focused on streaming data deduplication. When ingesting in, there are two things to look out for from a deduplication perspective - number 1 is to make sure there is no duplication on a batch level - i.e. batch 1 and batch 2 both have unique records, which can be done when reading the stream from bronze to silver, and then number 2 is to make sure there is no duplication at a table level, i.e. only insert a record if there is no duplicate record that already exists in the table.